display active filters as filter tiles
