+++ apps/projects/app/components/Shared/FilterBar/FilterBar.js
@@ -11,8 +11,8 @@ import {
 import Overflow from './Overflow'
 import FilterButton from './FilterButton'
 import FilterDropDown from './FilterDropDown'
+import prepareFilters from './prepareFilters'
 import { IconBigArrowDown, IconBigArrowUp } from '../../Shared'
-import { BOUNTY_STATUS } from '../../../utils/bounty-status'
 
 const StyledFilterBar = styled.div`
   width: 100%;
@@ -71,89 +71,6 @@ class FilterBar extends React.Component {
     this.props.handleFiltering(filters)
   }
 
-  /*
-    prepareFilters builds data structure for displaying filterbar dropdowns
-    data comes from complete issues array, issuesFiltered is used for counters
-  */
-  prepareFilters = (issues, bountyIssues) => {
-    let filters = {
-      projects: {},
-      labels: {},
-      milestones: {},
-      deadlines: {},
-      experiences: {},
-      statuses: {},
-    }
-
-    filters.statuses['not-funded'] = {
-      name: BOUNTY_STATUS['not-funded'],
-      count: issues.length - bountyIssues.length,
-    }
-    bountyIssues.map(issue => {
-      if (issue.data.workStatus in filters.statuses) {
-        filters.statuses[issue.data.workStatus].count++
-      } else {
-        filters.statuses[issue.data.workStatus] = {
-          name: BOUNTY_STATUS[issue.data.workStatus],
-          count: 1,
-        }
-      }
-    })
-
-    issues.map(issue => {
-      if (issue.milestone) {
-        if (issue.milestone.id in filters.milestones) {
-          filters.milestones[issue.milestone.id].count++
-        } else {
-          filters.milestones[issue.milestone.id] = {
-            ...issue.milestone,
-            count: 1,
-          }
-        }
-      } else {
-        if ('milestoneless' in filters.milestones) {
-          filters.milestones['milestoneless'].count++
-        } else {
-          filters.milestones['milestoneless'] = {
-            title: 'Issues without milestones',
-            id: 'milestoneless',
-            count: 1,
-          }
-        }
-      }
-
-      if (issue.labels.totalCount) {
-        issue.labels.edges.map(label => {
-          if (label.node.id in filters.labels) {
-            filters.labels[label.node.id].count++
-          } else {
-            filters.labels[label.node.id] = { ...label.node, count: 1 }
-          }
-        })
-      } else {
-        if ('labelless' in filters.labels) {
-          filters.labels['labelless'].count++
-        } else {
-          filters.labels['labelless'] = {
-            name: 'Issues without labels',
-            id: 'labelless',
-            count: 1,
-          }
-        }
-      }
-      // TODO: shouldn't it be reporitory.id?
-      if (issue.repository.id in filters.projects) {
-        filters.projects[issue.repository.id].count++
-      } else {
-        filters.projects[issue.repository.id] = {
-          name: issue.repository.name,
-          count: 1,
-        }
-      }
-    })
-    return filters
-  }
-
   generateSort = what => () => {
     const sortBy = this.state.sortBy
     sortBy.map(s => {
@@ -169,7 +86,7 @@ class FilterBar extends React.Component {
     const { handleSelectAll, allSelected, issues, bountyIssues, filters } = this.props
     // filters contain information about active filters (checked checkboxes)
     // filtersData is about displayed checkboxes
-    const filtersData = this.prepareFilters(issues, bountyIssues)
+    const filtersData = prepareFilters(issues, bountyIssues)
     return (
       <StyledFilterBar>
         <FilterButton>
