+++ apps/projects/app/components/Shared/FilterBar/index.js
@@ -1,1 +1,2 @@
 export { default } from './FilterBar'
+export { default as prepareFilters } from './prepareFilters'
